Refactor kernel tests
